refactor(server): rename file
