+++ server/src/parser/analyzer/utils/lookups.ts
